requirements.txt
inky==0.0.5
Pillow==7.0.0
pip==20.0.2
pylint==2.4.4
requests==2.23.0
setuptools==45.2.0
wheel==0.34.2
